requirement.txt
Django==1.11.15
django-filter==2.0.0
djangorestframework==3.8.2
ipaddress==1.0.22
Markdown==2.6.11
MySQL-python==1.2.5
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
